Move Order class of file.
